feat(text-tracks): Changed file references, some code improvements
This is a huge commit. Classes relating to the track API that were
originally kept in `src/js/track` have been moved to
`src/js/core/track`. This makes sense in a way, considering Moovie
polyfills the text track API, which is a **core** feature. Amongst
the move I got a little carried away and refactored a bit as well.